fix(preprocessor): use absolute paths
According to the spec, the absolute file path should be used in the
coverage report. Currently karma-coverage substitutes this with a
path relative to `baseDir`. This causes problems when trying to
combine coverage with other reports not generated by karma-coverage.
Fixing this would help with #1.

coverag.json spec for reference:
 - https://github.com/gotwarlost/istanbul/blob/master/coverage.json.md