Update istanbul to support ES6 and async/await
The istanbul supports ES6 and async/await (ES7) in it's alpha version but karma-coverage does not support this so we can see parsing error from some syntax such as async/await , comma dangle in function or destructed objects with default in parameter ..

https://github.com/gotwarlost/istanbul/tree/v1.1.0-alpha.1
Although the version is manually updated, it is surely not working with karma-coverage. Do you have plan to support the version?