fix(preprocessor): merging source maps
This commit fixes an issue where karma would be unable to report
correctly mapped stacktraces back to original sources. Instead of
manually merging source maps, the default instrumenter, istanbul,
supports passing in an input source map and having that merged
automatically with the internal source map of instrumented code.

The merged source map is now available for access using
instrumenter.lastSourceMap(), so that is what file.sourceMap gets
updated to (if the merge completed successfully).

Istanbul reporting still requires the original source map only, so the
check to register the source map with the source store is made before
this merged source map is used. This allows istanbul to use the original
incoming source map, while still allowing karma to use the merged source
map.